+++ spotifywebapipython/sautils.py
@@ -56,6 +56,37 @@ def export(fn):
     return fn
     
 
+def validateDelay(
+        self, 
+        delay:float, 
+        default:float=0.5, 
+        maxDelay:float=10
+        ) -> float:
+    """
+    Validates a delay value.
+        
+    Args:
+        delay (int):
+            The delay value to validate.
+        default (int):
+            The default delay value to set if the user-input delay is not valid.
+        maxDelay (int):
+            The maximum delay value allowed.  
+            Default is 10.
+    """
+    if (isinstance(delay,int)): 
+        delay = float(delay)
+            
+    if (not isinstance(delay,float)) or (delay < 0): 
+        result = default
+    elif (delay > maxDelay): 
+        result = maxDelay
+    else:
+        result = delay
+            
+    return result
+
+
 def _xmlGetInnerText(elmNode, default:str=None) -> str:
     """
     Finds the specified xml node inner text from the root Element object and returns its value